Refactor and optimize report services
Updated `FrozenReportService` to remove unnecessary casting for `EnrollmentId`. Refactored `ProfitSharingSummaryReportService` to improve readability, performance, and maintainability:
- Initialized `_hoursThreshold` inline and removed static constructor.
- Simplified `lineItem.NumberOfMembers` calculation using `Count()` directly.
- Removed `fiscalBeginDate` parameter from `GetReportDetailFilterExpression`.
- Made LINQ query asynchronous with `ToListAsync`.
- Removed explicit cast for `Balance` property.
- Updated XML documentation to reflect method signature changes.

+++ src/services/src/Demoulas.ProfitSharing.Services/Reports/FrozenReportService.cs
@@ -940,7 +940,7 @@ public class FrozenReportService : IFrozenReportService
                         Loans = x.Loans ?? 0,
                         ProfitSharingAmount = x.ProfitSharingAmount ?? 0,
                         GrossWages = x.GrossWages,
-                        EnrollmentId = (byte)(x.EnrollmentId ?? 0),
+                        EnrollmentId = (x.EnrollmentId ?? 0),
                         IsExecutive = x.PayFrequencyId == PayFrequency.Constants.Monthly,
                     }).ToList(),
                     Total = pagedData.Total
